requirements.txt
openai>=1.3.0
python-dotenv
fpdf2>=2.7.6
Pillow
